requirement.txt
kafka-python
pandas
boto3